backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
requests
python-docx
PyPDF2
